// src/interface.d.ts
// Global types
export type LabelColor = 'primary' | 'secondary' | 'tertiary' | 'destructive' | 'brand-primary-blue' | 'brand-primary-green' | 'brand-primary-red' | 'neutral-100' | 'custom-brand-accent';
export type BackgroundColor = 'primary' | 'secondary' | 'tertiary';

export type ActionType = 'primary' | 'secondary' | 'danger';

export type Sizes = 'xxs' | 'xs' | 's' | 'm' | 'l' | 'xl' | 'xxl';